Added BSWAP* byte swapping macros to util.h.  These trigger an annoying warning so disabled that in the Makefile.
Added support for automatically byte swapping ephemeris files (untested).
Calculate blocksize using the ephemeris file header, eliminating the need
for the hardcoded table.  Relies on the Fortran direct access output zero
padding the file, which seems to be the case for the ones I have.  Should
check if this is true in general.